+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 - Fixed a bug with certain types of keyboards on macOS conflicting with internal CodeMirror commands.
 - Prevent opening of a user's home directory.
 - The citation rendering plugin won't render the domain parts of Emails anymore.
+- Markdown links in braces won't include the closing brace anymore.
 
 ## Under the Hood